FIX: global min check on invariant calculations in binary and ternary mapping (#639)
This addresses part of #638 to fix tie-lines not being computed out of a three-phase tie-triangle. This fix checks for degenerate composition sets when performing a global min check on computed invariants. Without this check, it was possible that a tie-triangle at global min was falsely rejected as a metastable equilibria.

This also seems to fix some of the troublesome binaries that were brought up in #564:
- [AuSn-13Don](https://phasediagrams.org/phase-diagram/AuSn-13Don.tdb)
- [BiLi-13Liu](https://phasediagrams.org/phase-diagram/BiLi-13Liu.tdb)
- [PdTi-11Guo](https://phasediagrams.org/phase-diagram/PdTi-11Guo.tdb)